+++ examples/styled-components-static-rendering/src/index.js
@@ -54,7 +54,7 @@ register({
 
     // Start react.
     renderer(
-      <Router navigation={navigation}>
+      <Router hashScrollBehavior='smooth' navigation={navigation}>
         <App />
       </Router>,
       document.getElementById("root")