+++ packages/react-navi/src/index.tsx
@@ -1,3 +1,4 @@
+export { HashScroll } from './HashScroll'
 export { Link } from './Link'
 export { NaviProvider } from './NaviProvider'
 export { NotFoundBoundary } from './NotFoundBoundary'